pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "antspynet"
version = "0.3.3"
description = "A collection of deep learning architectures ported to the python language and tools for basic medical image processing."
readme = "README.md"
authors = [
    {name = "Nicholas J. Tustison", email = "ntustison@gmail.com"},
    {name = "Brian B. Avants"},
    {name = "Nick Cullen"}
]
license = {text = "Apache License 2.0"}
keywords = ["ANTsPyNet", "deep learning", "medical image processing"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "tensorflow",
    "antspyx>=0.6.1",
    "scikit-learn",
    "requests",
    "statsmodels",
    "matplotlib"
]

[project.urls]
homepage = "https://github.com/ANTsX/ANTsPyNet"

[tool.setuptools.packages]
find = { include = ["antspynet","antspynet.*"] }

[tool.setuptools]
license-files = []

[project.scripts]
antsMouseBrainExtraction = "antspynet.cli.mouse_brain_extraction:main"

